+++ src/main/nextflow/co2footprint/FileCreation/ProvenanceFileCreator.groovy
@@ -7,11 +7,14 @@ import groovyx.gpars.agent.Agent
 import nextflow.co2footprint.Config.ProvenanceFileConfig
 import nextflow.co2footprint.Records.CO2Record
 import nextflow.co2footprint.Records.CO2RecordTree
+import nextflow.co2footprint.Records.CiRecordCollector
 import nextflow.trace.TraceHelper
 
 import java.nio.file.Path
 import java.time.Duration
 import java.time.Instant
+import java.time.LocalDateTime
+import java.time.ZoneOffset
 
 @Slf4j
 class ProvenanceFileCreator extends BaseFileCreator {
@@ -36,7 +39,7 @@ class ProvenanceFileCreator extends BaseFileCreator {
 
         if(!config.enabled) {
             this.metaClass.create = { -> null }
-            this.metaClass.write = { CO2RecordTree X -> null }
+            this.metaClass.write = { CO2RecordTree X, CiRecordCollector Y -> null }
             this.metaClass.close = {  -> null }
         }
     }
@@ -56,14 +59,41 @@ class ProvenanceFileCreator extends BaseFileCreator {
      *
      * @param co2RecordTree A hierarchically structured record tree
      */
-    void write(CO2RecordTree co2RecordTree) {
-        Map co2TreeMap = transformToJsonLd(co2RecordTree.toMap(emissionMetricsOnly, false, false))
+    void write(CO2RecordTree co2RecordTree, CiRecordCollector timeCiRecordCollector=null) {
+        if (timeCiRecordCollector?.timeCIs) {
+            co2RecordTree.metaData.put(
+                'carbon_intensity_records',
+                ciRecordsToJsonLd(timeCiRecordCollector.timeCIs)
+            )
+        }
+        Map co2TreeMap = transformToJsonLd(
+                co2RecordTree.toMap(emissionMetricsOnly, false, false)
+        )
         JsonBuilder jsonBuilder = new JsonBuilder(co2TreeMap)
 
         dataWriter = new Agent<PrintWriter>(file)
 
         file.print(jsonBuilder.toPrettyString())
         file.flush()
+    }
+
+    /**
+     * Transform the carbon intensity records to a valid JSON-LD file string.
+     * 
+     * @param ciRecords Carbon intensity records as a Map
+     * @return
+     */
+    private static Map<String, Object> ciRecordsToJsonLd(Map<Instant, Number> ciRecords ) {
+        Map<String, Object> ciRecordsMap = ['@type': 'schema:ItemList']
+        ciRecords.eachWithIndex { Instant instant, Number ci, Integer i ->
+            ciRecordsMap.put( 'itemListElement', 
+                [ '@type': 'schema:ListItem', position: i, item: [
+                    '@type': 'schema:Observation', observationDate: instant.toString(), value: ci, unitText: 'g/kWh'
+                    ]
+                ]
+            )
+        }
+        return ciRecordsMap
     }
 
     /**
@@ -189,6 +219,10 @@ class ProvenanceFileCreator extends BaseFileCreator {
         }
 
         return ldMap
+    }
+    
+    private Map<String, Object> addCo2RecordsToMap(Map<String, Object> co2Records, Map<String, Object> treeMap) {
+        
     }
 
     static CO2RecordTree read(Path path) {